requirements.txt
pandas
sqlalchemy
psycopg2-binary
flask
pyyaml
matplotlib
seaborn
os
